--- client/src/app/layout.tsx
@@ -1,30 +1,0 @@
-import Navbar from "@/components/layout/Navbar";
-import "./globals.css";
-import { Nunito } from "next/font/google";
-
-const nunito = Nunito({
-  subsets: ["latin"],
-  weight: ["300", "400", "500", "600"],
-});
-
-export const metadata = {
-  title: "App",
-  description: "Generated by create next app",
-};
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-      <html lang="en" className="w-full h-full">
-        <body className={`${nunito.className} ${"bg-ivory"}`}>
-          <div>{children}</div>
-          <div className="fixed bottom-0">
-            <Navbar />
-          </div>
-        </body>
-      </html>
-  );
-}
