--- client/src/app/service/create/hangout/page.tsx
@@ -1,24 +1,0 @@
-import { Text } from "@/components/elements/Text";
-import { Title } from "@/components/elements/Title";
-import { HangoutForm } from "@/components/service/Create/Hangout/hangout-form";
-import React from "react";
-
-const Hangout = () => {
-  return (
-    <div className="p-10 py-20">
-      <Title fontsize="text-[24px]">
-        したいことを
-        <br />
-        登録しましょう！
-      </Title>
-      <div className="py-4">
-        <Text>ooさんは何をして遊びたいですか？</Text>
-      </div>
-      <div>
-        <HangoutForm />
-      </div>
-    </div>
-  );
-};
-
-export default Hangout;
