+++ client/src/components/layout/Layout.tsx
@@ -1,0 +1,26 @@
+import Navbar from "@/components/layout/Navbar";
+// import "@/pages/globals.css";
+import { Nunito } from "next/font/google";
+
+const nunito = Nunito({
+  subsets: ["latin"],
+  weight: ["300", "400", "500", "600"],
+});
+
+export const metadata = {
+  title: "App",
+  description: "Generated by create next app",
+};
+
+export const RootLayout = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <div lang="en" className="w-screen h-screen">
+      <div className={`${nunito.className} ${"bg-ivory"}`}>
+        <div>{children}</div>
+        <div className="fixed bottom-0">
+          <Navbar />
+        </div>
+      </div>
+    </div>
+  );
+};
