+++ client/src/pages/api/auth/[...nextauth].ts
