+++ client/src/pages/favicon.ico
